- Modifies runtime script to cater for Microsoft users - Adds Emetic and Lethal posion to the map - Updates Chinese Translations from suanjiansalt - Makes Blade bodyguard wait 3mins before searching for dead target

+++ script/CopyHantu.ts
@@ -14,7 +14,7 @@ export const afterDeploy: ModScript["afterDeploy"] = async function (context, ap
   fs.copyFileSync(context.config.runtimePath + '/chunk13.rpkg', context.config.runtimePath + '/chunk9patch200.rpkg');
   fs.copyFileSync(context.config.runtimePath + '/chunk13patch1.rpkg', context.config.runtimePath + '/chunk9patch201.rpkg');
   fs.copyFileSync(context.config.runtimePath + '/chunk13patch2.rpkg', context.config.runtimePath + '/chunk9patch202.rpkg');
-  fs.copyFileSync(context.config.runtimePath + '/chunk13patch3.rpkg', context.config.runtimePath + '/chunk9patch203.rpkg');
+  if(fs.existsSync(`${context.config.runtimePath}/chunk13patch3.rpkg`)) fs.copyFileSync(context.config.runtimePath + '/chunk13patch3.rpkg', context.config.runtimePath + '/chunk9patch203.rpkg');
   
   return Promise.resolve();
 };
